requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
attrs==25.3.0
b2sdk==2.8.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
frozenlist==1.5.0
gunicorn==22.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
logfury==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
packaging==24.2
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
razorpay==1.4.2
reportlab==4.4.1
requests==2.32.3
SQLAlchemy==2.0.41
twilio==9.5.2
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.19.0
